requirements.txt
aiohttp==3.9.4
matplotlib==3.8.4
openpyxl==3.1.2
pandas==2.0.2
pydantic==1.10.8
python-dotenv==1.0.1
scikit_learn==1.2.2
